Pull request #3720: src: fix config parsing issues seen on 32bit systems
Merge in SNORT/snort3 from ~ASERBENI/snort3:32bit_issue to master

Squashed commit of the following:

commit 8137a4fca03573398a89f011ce3e66743b9c4154
Author: Andrii Serbeniuk <aserbeni@cisco.com>
Date:   Wed Jan 4 17:15:04 2023 +0200

    src: address numbers parsing related concerns

commit 99895d8af9eb73b5646d54dc063322b910e467ea
Author: Andrii Serbeniuk <aserbeni@cisco.com>
Date:   Wed Jan 4 15:35:20 2023 +0200

    framework: add strtoul methods to Value class

commit 8e431851bc5416cb845684108d5a3c0a2407ecc3
Author: Andrii Serbeniuk <aserbeni@cisco.com>
Date:   Wed Dec 21 16:48:01 2022 +0200

    framework: change range check types to int64_t

    long may not be enough on 32bit platforms, where it's only 4 bytes long. issue initially found with seq ips option, where a valid value of 3,927,875,496 would be perceived as erroneous because it would not fit in 4 byte signed long (max value is 2,147,483,647)

commit 59fec3d494cc8560754123688c4d9de7e216bbee
Author: Andrii Serbeniuk <aserbeni@cisco.com>
Date:   Tue Dec 20 14:38:30 2022 -0500

    dce_rpc: add errno resets during uuid parsing

+++ src/managers/ips_manager.cc
@@ -130,10 +130,10 @@ static bool set_arg(
         if ( p->is_wild_card() )
             val = opt;
 
-        long n = (long)strtoll(val, &end, 0);
+        int64_t n = strtoll(val, &end, 0);
 
         if ( !*end )
-            v.set(n);
+            v.set((double)n);
         else
             ok = false;
     }
